# tweak f64 deserialization to make float roundtrips work