This is a new attempt at fixing issue #536 using the latest version of @Alexhuszagh 's `minimal-lexical` library for deserialization of f64 values. The goal is to make float roundtrip work without severely impacting compile times.

I have:
- checked that the crate still compiles with Rust 1.31
- added a sample of roundtrip tests that fail with the current deserializer but pass with this branch (also did some testing with `quickcheck` but didn't add that here as it would bring it a new dev-dependency)
- ran [this loop test](https://github.com/serde-rs/json/issues/536#issuecomment-583708730) for over 45 minutes
- measured changes in compile time using `perf` (see table below)
- benchmarked this branch and compared it against master (results at the bottom)

Note that:
- as used here (with default features off), the `minimal-lexical` crate itself does *not* have any dependencies
- but `minimal-lexical` will use a heap-allocated `Vec` for its internal, intermediate buffer instead of stack allocated one (which would bring in the `arrayvec` dependency)
- there's also the option of enabling the `arrayvec` feature or writing a custom stack allocated `Vec` that doesn't bring any extra dependencies -- though the heap-allocated `Vec` doesn't seem to degrade performance as measured by json-benchmark

## compile time changes

(compiler used: 1.43.1)

I have measured `perf stat cargo $command` on the `serde-json` crate after a `cargo clean` for these values of `$command`: `check`, `build`, `build --release` and a faster version of `build --release`. The table reports the changes in `instructions:u` both in absolute value and as a percentage. The table also includes the initial and the change in wall time, to get a more "human" feel for the change. I ran perf a few times for each command; the values reported here are median values. 

| Command         | Δ insn           | Time (master) | Δ Time           |
| --------------- | ---------------- | --------------|------------------|
| `check`         | +1.848G (+1.6%)  | 14.584s       | +0.232s (+1.59%) |
| `build`         | +3.442G (+2.75%) | 15.471        | +0.323s (+2.09%) |
| `b --release`   | +6.004G (+1.46%) | 30.432        | +0.675s (+2.219%)|
| `b --release`++ | +6.219 (+5.28%)  | 14.729        | +0.31s (+2.106%) |

The faster `build --release` consists of adding the profile override shown below to compile all build scripts and proc macros without optimizations.

``` toml
[profile.release.build-override]
codegen-units = 16
debug = false
debug-assertions = false
opt-level = 0
overflow-checks = false
```

(off-topic: Your mileage may vary though with this profile trick. I known that un-optimized `bindgen` runs very slowly so this actually backfires when you have that in your dependency graph (or at least that was the case many months ago; haven't use `bindgen` with embedded code in a long time))

## runtime perf

I ran `json-benchmark` with this branch and master and got these `parse` results. Again, all values are median values out of a few runs.

| Dataset      | DOM - master | DOM - PR | STRUCT - master | STRUCT - PR |
| ------------ | ------------ | ---------|-----------------|-------------|
| canada       |  250 MB/s    | 260 MB/s | 480 MB/s        | 480 MB/s
| citm_catalog |  380 MB/s    | 400 MB/s | 830 MB/s        | 810 MB/s
| twitter      |  280 MB/s    | 280 MB/s | 610 MB/s        | 620 MB/s

These values fluctuated quite a bit so I wouldn't say this PR improves or degrades performance.

@dtolnay do these changes in compile time, in exchange for the round-trip correctness, seem like an acceptable trade-off to you?