@rw 2020-05-13
WDYT about testing every f32? :-]