@japaric 2020-05-14
> errors: 1 with 0x95ae43fd
> errors: 2 with 0x15ae43fd

OK, I see the issue. when serializing we use the f32 variant of ryu to serialize f32 values and the f64 variant to deal with f64 inputs. when deserializing everything goes through the 64-bit variant of minimal-lexical and then output gets casted to an f32 if the user requested that; doing it like that results in an error of 1 ULP for `0x15ae43fd` (`0x95ae43fd` is the negated value).

using the 32-bit variant of minimal-lexical (on the ryu-stringified value of `0x15ae43fd`) does produce the correct value so we should use that one if the requested output is the `f32` type