@japaric 2020-05-14
> we should use that one if the requested output is the f32 type.

Done, f32 roundtrip is now working for all finite f32 values (and the program seems to terminate a bit faster?). This change will bring in the `f32` monophormization of `parse_float`; that seems to increase the # of instructions reported by `perf stat cargo build` by ~0.13%.