@rw 2020-05-15
@japaric The `serde_json::to_string` is probably allocating on every execution. I'd try re-using a `String` as a buffer, instead.