@japaric 2020-05-15
ah, yes. You are right. Deserialization also allocates because `Deserializer` contains a `Vec` (that was there before this PR). I don't see a way to re-use a `Deserializer` to avoid that `Vec` being allocated and deallocated each time a value is deserialized though, at least not using the public API.

Using `to_writer`, testing the whole 32-bit range takes ~11 minutes in single-threaded mode, ~5 minutes with 8 threads (using just `std::thread` and no external dependencies) and ~6 minutes with 4 threads -- I guess more threads don't help in the program I wrote because there's no load balancing. I can commit either version and if the test is deemed to take too long it could be run on a daily basis using cron (rather than on each commit / PR).