@rw 2020-05-15
@japaric Could you post your updated "zero-alloc in the hot sections" version?