@japaric 2020-05-26
@dtolnay did you have a chance to look at this PR? :eyes: