@dtolnay 2020-06-07
Courtesy of libFuzzer, here is an input that triggers the out of bounds write:

```rust
fn main() {
    let j = "144440.000312571602869014401444400000312571602869014031257160286901444400000312144440000031257160286901444444440000312571602869014444000003125716028690144440000031257160286901444400000312571602869014444000031250312571602312571602860003125716023125716028694400003125716028690144440000031254400000312571602869014444000003125716028690144440000312503125716023125716028600031257160231257160286";
    let _ = serde_json::from_str::<f64>(j);
}
```

```console
==41341== Invalid write of size 8
==41341==    at 0x4841C24: memmove (in /usr/lib/x86_64-linux-gnu/valgrind/vgpreload_memcheck-amd64-linux.so)
==41341==    by 0x123B6C: core::intrinsics::copy (intrinsics.rs:2127)
==41341==    by 0x11E02F: minimal_lexical::math::insert_many (math.rs:278)
==41341==    by 0x12EEBF: minimal_lexical::math::small::ishl_limbs (math.rs:664)
==41341==    by 0x12F154: minimal_lexical::math::small::ishl (math.rs:678)
==41341==    by 0x12CA89: minimal_lexical::math::Math::ishl (math.rs:1046)
==41341==    by 0x12CC2D: minimal_lexical::math::Math::imul_pow2 (math.rs:1025)
==41341==    by 0x127626: minimal_lexical::bhcomp::small_atof (bhcomp.rs:176)
==41341==    by 0x128778: minimal_lexical::bhcomp::bhcomp (bhcomp.rs:225)
==41341==    by 0x121DC0: minimal_lexical::algorithm::fallback_path (algorithm.rs:176)
==41341==    by 0x11BF6B: minimal_lexical::parse::parse_float (parse.rs:80)
==41341==    by 0x11657A: serde_json::de::Deserializer<R>::f64_from_parts (de.rs:752)
==41341==  Address 0x4ab8430 is 0 bytes after a block of size 160 alloc'd
==41341==    at 0x483A7F3: malloc (in /usr/lib/x86_64-linux-gnu/valgrind/vgpreload_memcheck-amd64-linux.so)
==41341==    by 0x13790B: alloc::alloc::alloc (alloc.rs:80)
==41341==    by 0x137E23: <alloc::alloc::Global as core::alloc::AllocRef>::alloc (alloc.rs:174)
==41341==    by 0x133E50: alloc::raw_vec::finish_grow (raw_vec.rs:514)
==41341==    by 0x1348DA: alloc::raw_vec::RawVec<T,A>::grow_amortized (raw_vec.rs:440)
==41341==    by 0x134346: alloc::raw_vec::RawVec<T,A>::try_reserve (raw_vec.rs:281)
==41341==    by 0x134B22: alloc::raw_vec::RawVec<T,A>::reserve (raw_vec.rs:267)
==41341==    by 0x130998: alloc::vec::Vec<T>::reserve (vec.rs:505)
==41341==    by 0x1387B3: minimal_lexical::math::reserve (math.rs:336)
==41341==    by 0x1323E0: <minimal_lexical::bignum::Bigint as core::default::Default>::default (bignum.rs:16)
==41341==    by 0x12CB30: minimal_lexical::math::Math::from_u64 (math.rs:991)
==41341==    by 0x12752D: minimal_lexical::bhcomp::small_atof (bhcomp.rs:145)
```