@rw 2020-06-12
@dtolnay Thanks! Should we create a new issue to dig into the `unsafe` usage in minimal-lexical?